requirements.txt
rembg
pillow
fastapi
uvicorn
shazamio
python-multipart
